requirements.txt
asgiref==3.3.1
ddt==1.4.2
Django==3.1.7
django-braces==1.14.0
django-model-utils==4.1.1
django-star-ratings==0.9.1
Pillow==8.1.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
